==> Tracert_Experiments/Single run experiment/7_alltraceroute_dist_arima.ipynb <==
# https://stackoverflow.com/questions/21321670/how-to-change-fonts-in-matplotlib-python
# https://towardsdatascience.com/ways-to-evaluate-regression-models-77a3ff45ba70
# https://www.kaggle.com/pmarcelino/comprehensive-data-exploration-with-python

import pandas as pd
import numpy  as np
import json
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings("ignore",category=DeprecationWarning)
warnings.simplefilter("ignore")
hfont = {'fontname':'Monaco'}
hfont = {'fontsize':'24'}

# Making the plots standard 
%matplotlib inline
plt.rcParams["figure.figsize"] = [16, 5]

#import the training and test pickles 
train_df = pd.read_pickle('train_df.pickle')
test_df = pd.read_pickle('test_df.pickle')

#get the training std 
trainrtt_mean = train_df['last_rtt'].mean()
trainrtt_std = train_df['last_rtt'].std()

print(trainrtt_mean,trainrtt_std)

# Choose your features and target
X_train = train_df['normalizzed_distance'].values.reshape(-1,1)
y_train = train_df['normalizzed_rtt'].values
        
X_test = test_df['normalizzed_distance'].values.reshape(-1,1)
y_test = test_df['normalizzed_rtt'].values



#Importing the module
import pmdarima as pm

#first model 
arima_model = pm.auto_arima(y_train,test="adf",trace=True,error_action='ignore',supress_warnings = True)

#Summary of the model
arima_model.summary()

import statsmodels.api as sm

# Get the best order
order = arima_model.order
model = sm.tsa.ARIMA(y_test, order=order)
arima_model = model.fit()

# Make predictions on the test set
arima_predictions = arima_model.fittedvalues

# calculate the mean squared error and standard deviation
from sklearn.metrics import mean_squared_error
rmse_arima =  np.sqrt(mean_squared_error(arima_predictions,y_test))
drmse_arima = rmse_arima * trainrtt_std
print(f"Root Mean Squared Error and denormalised rmse: {rmse_arima,drmse_arima}")
