==> Tracert_Experiments/Single run experiment/10_alltraceroute_allfeatures_train.ipynb <==
# Load the libraries
import pandas as pd
import numpy  as np
import json
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn as sk
import warnings
import platform
import sys
warnings.filterwarnings("ignore",category=DeprecationWarning)
warnings.simplefilter("ignore")

# Making the plots standard 
%matplotlib inline
plt.rcParams["figure.figsize"] = [16, 5]

#Hardware and software details 
print(f"Python Platform: {platform.platform()}")
print(f"Python {sys.version}")
print(f"Pandas {pd.__version__}")
print(f"Scikit-Learn {sk.__version__}")

# import the pickled dataframes you will be using for the experiments
latency_df = pd.read_pickle('latency_withprobe.pickle')

#remove the null values from the experiment_df 
nan_rows = latency_df[latency_df['last_rtt'].isnull()]
latency_df = latency_df.dropna(subset=['last_rtt'])

len(nan_rows)

#remove indexes where last_rtt is greater than 100ms 
latency_df = latency_df[latency_df['last_rtt'] < 100]
len(latency_df)

model_df = latency_df.copy()
model_df.columns

#Before applying models randomise selection of training and test data
import itertools
import random

test_indices = []
train_indices = []
        
array1 = model_df['prb_id'].unique()
array2 = model_df['dst_id'].unique()

# Creating all possible pairs
pairs = list(itertools.product(array1, array2))

# Randomly selecting 10 pairs
selected_pairs = random.sample(pairs, 10)

# Removing selected pairs from the original list
for pair in selected_pairs:
    pairs.remove(pair)

# Creating separate lists
selected_list = selected_pairs
remaining_list = pairs

train_dfs = []
for i,k in remaining_list:
    temp_df = model_df.loc[(model_df['prb_id'] == i) & (model_df['dst_id'] == k)]
            
    # Append the piece to the selected data
    train_dfs.append(temp_df)
            
    train_df = pd.concat(train_dfs)
        
test_dfs = []
for i,k in selected_list:
    temp_df = model_df.loc[(model_df['prb_id'] == i) & (model_df['dst_id'] == k)]
            
    # Append the piece to the selected data
    test_dfs.append(temp_df)

    test_df = pd.concat(train_dfs)

selected_pairs

remaining_list

#normalise the rtt values 
trainrtt_mean = train_df['last_rtt'].mean()
trainrtt_std = train_df['last_rtt'].std()
train_df['normalizzed_rtt'] = (train_df['last_rtt'] - trainrtt_mean) / trainrtt_std

# normalize the 'distance' column
traindist_mean = train_df['distance'].mean()
traindist_std = train_df['distance'].std()
train_df['normalizzed_distance'] = (train_df['distance'] - traindist_mean) / traindist_std
        
#normalise the test rtt values 
test_df['normalizzed_rtt'] = (test_df['last_rtt'] - trainrtt_mean) / trainrtt_std

# normalize the 'distance' column for the test set 
test_df['normalizzed_distance'] = (test_df['distance'] - traindist_mean) / traindist_std


#check for nan values in normalised rtt and distance
nan_train = train_df[train_df['normalizzed_rtt'].isnull()]
nan_test = test_df[test_df['normalizzed_rtt'].isnull()]
print(len(nan_train),len(nan_test))

#pickle the train and test dataframes
train_df.to_pickle('train_df.pickle')
test_df.to_pickle('test_df.pickle')
